+++ src/App/backend/test/Altinn.App.Api.Tests/Features/Bootstrap/FormBootstrapServiceTests.cs
@@ -840,12 +840,7 @@ public class FormBootstrapServiceTests
                     It.IsAny<CancellationToken>()
                 )
             )
-            .ThrowsAsync(
-                new PlatformHttpException(
-                    new PlatformHttpResponse { StatusCode = HttpStatusCode.Forbidden },
-                    "Forbidden"
-                )
-            );
+            .ThrowsAsync(new PlatformHttpException(HttpStatusCode.Forbidden, "Forbidden"));
 
         var service = CreateService();